address comments nad update libs
