+++ .github/workflows/android-perf.yml
@@ -39,7 +39,7 @@ jobs:
     - uses: actions/checkout@v3
 
     - name: Setup Node.js environment
-      uses: actions/setup-node@v2.4.0
+      uses: actions/setup-node@v3
 
     - name: Install AppCenter CLI
       run: npm install -g appcenter-cli
