+++ binding/android/eagle/proguard-rules.pro
@@ -1,2 +1,6 @@
 -keep class ai.picovoice.eagle.*Exception { <init>(...); }
 -keep class ai.picovoice.eagle.Eagle
+-keep class ai.picovoice.eagle.EagleProfile
+-keep class ai.picovoice.eagle.EagleProfiler
+-keep class ai.picovoice.eagle.EagleProfilerEnrollResult
+-keep class ai.picovoice.eagle.EagleProfilerEnrollFeedback
