+++ demo/android/EagleDemo/build.gradle
@@ -17,7 +17,7 @@ allprojects {
         google()
         mavenCentral()
         maven {
-            url 'https://s01.oss.sonatype.org/content/repositories/aipicovoice-1235/'
+            url 'https://s01.oss.sonatype.org/content/repositories/aipicovoice-1237/'
         }
     }
 }
